@@ -1,6 +1,6 @@
 package com.example.ben.smarttext;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 public class Contact {
 
