@@ -3,7 +3,7 @@ package com.example.ben.smarttext;
 import android.app.Service;
 import android.content.Intent;
 import android.os.IBinder;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import android.content.Context;
 import android.util.Log;
