@@ -1,6 +1,6 @@
 package com.example.ben.smarttext;
 
-import android.arch.persistence.room.TypeConverter;
+import androidx.room.TypeConverter;
 
 import java.util.Date;
 
